fix: 신규 매물 dynamic import 적용
