+++ src/components/main/section-skeleton/SectionSkeleton.tsx
@@ -1,0 +1,7 @@
+import { Skeleton } from '@mui/material';
+
+const SectionSkeleton = () => {
+  return <Skeleton sx={{ my: 6, mx: 2, width: '100%', height: '300px' }} />;
+};
+
+export default SectionSkeleton;
